requirements.txt
absl-py==1.4.0
altair==4.2.2
astunparse==1.6.3
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==4.4.2
deepface==0.0.79
entrypoints==0.4
extra-streamlit-components==0.1.56
filelock==3.12.0
Flask==2.2.3
flatbuffers==1.12
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
idna==3.4
imageio==2.28.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
ml-dtypes==0.1.0
moviepy==1.0.3
mtcnn==0.1.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
plotly==5.14.1
proglog==0.1.10
protobuf==3.19.6
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.6.0
Pympler==1.0.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
retina-face==0.0.13
rich==13.3.4
rsa==4.9
scipy==1.10.1
six==1.15.0
smmap==5.0.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
streamlit==1.21.0
streamlit-authenticator==0.2.1
streamlit-option-menu==0.3.2
tenacity==8.2.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.0
tensorflow-metal==0.5.0
termcolor==2.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.15.0
